Create a HostStatePartial and a new HostManager classes
In order to track the resources consumed by the preemptible instances we
introduce the HostStatePartial class inheriting from the original
HostStace class. It is able to return two different host states:

 - A partial state, taking into account only normal instances (i.e.
   considering the resources consumed by the preemptible instances as
   free).
 - A full state, taking into account _all_ instances inside a host.

The new HostManager class extends the original manager (that is
nova.scheduler.host_manager.HostManager), adding a new method for
obtaining the host partial states.